requirements.txt
django
datetime
python-dateutil
django-cors-headers